# task3-RESTful_API/models/auth.go
package models

import (
	"time"

	"golang.org/x/crypto/bcrypt"
)

// User represents a user in the database
type User struct {
	ID        uint      `json:"id" gorm:"primaryKey;autoIncrement"`
	Username  string    `json:"username" gorm:"uniqueIndex;not null"`
	Password  string    `json:"-" gorm:"not null"` // "-" means don't include in JSON
	CreatedAt time.Time `json:"created_at" gorm:"autoCreateTime"`
	UpdatedAt time.Time `json:"updated_at" gorm:"autoUpdateTime"`
}

// SetPassword hashes and sets the user password
func (u *User) SetPassword(password string) error {
	hashedPassword, err := bcrypt.GenerateFromPassword([]byte(password), bcrypt.DefaultCost)
	if err != nil {
		return err
	}
	u.Password = string(hashedPassword)
	return nil
}

// CheckPassword verifies if the provided password matches the user's password
func (u *User) CheckPassword(password string) bool {
	err := bcrypt.CompareHashAndPassword([]byte(u.Password), []byte(password))
	return err == nil
}

// Session represents an authentication session/token
type Session struct {
	ID        uint      `json:"id" gorm:"primaryKey;autoIncrement"`
	Token     string    `json:"token" gorm:"uniqueIndex;not null"`
	Username  string    `json:"username" gorm:"not null;index"`
	CreatedAt time.Time `json:"created_at" gorm:"autoCreateTime"`
	ExpiresAt time.Time `json:"expires_at" gorm:"index"`
}

// Config represents the application configuration (for YAML loading)
type Config struct {
	Users []User `yaml:"users"`
}

// LoginRequest represents the login request payload
// @Description Login credentials
type LoginRequest struct {
	Username string `json:"username" example:"admin"`
	Password string `json:"password" example:"admin123"`
}

// LoginResponse represents the login response
// @Description Login response with access token
type LoginResponse struct {
	Token   string `json:"token" example:"a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6"`
	Message string `json:"message" example:"Login successful"`
}

// LogoutRequest represents the logout request payload
type LogoutRequest struct {
	Token string `json:"token"`
}

// Validate validates the login request
func (r *LoginRequest) Validate() error {
	if r.Username == "" {
		return NewValidationError("username is required")
	}
	if r.Password == "" {
		return NewValidationError("password is required")
	}
	return nil
}